Refactor Admin page
