--- src/components/AddUser/index.ts
@@ -1,1 +1,0 @@
-export * from "./AddUser";
