--- src/components/ApproveHours/ApproveHours.tsx
@@ -1,14 +1,0 @@
-import { Button, Card, Typography } from "@mui/material";
-
-export const ApproveHours: React.FC = () => {
-  return (
-    <Card>
-      <Typography variant="h6">Approve Hours</Typography>
-      <Button variant="contained">
-        Approve
-      </Button>
-    </Card>
-  );
-};
-
-export default ApproveHours;
