--- src/components/ApproveHours/index.ts
@@ -1,1 +1,0 @@
-export * from "./ApproveHours";
