+++ src/components/index.ts
@@ -1,13 +1,11 @@
 export * from "./AddEventDrawer";
 export * from "./AddShifts";
-export * from "./AddUser";
-export * from "./ApproveHours";
+export * from "./ApproveHoursDrawer";
 export * from "./BigStatistic";
 export * from "./ContextMenu";
 export * from "./Drawer";
 export * from "./EventDrawer";
 export * from "./EventListing";
-export * from "./GenerateReport";
 export * from "./HourDrawer";
 export * from "./Login";
 export * from "./NavBar";
