+++ src/pages/Admin/Admin.styles.tsx
@@ -1,5 +1,7 @@
+import { theme } from "@/styles";
 import styled from "@emotion/styled";
-import { Box } from "@mui/material";
+import { Box, Card } from "@mui/material";
+import { Link } from "react-router-dom";
 
 export const StyledPage = styled(Box)({
   display: "inline-flex",
@@ -7,6 +9,7 @@ export const StyledPage = styled(Box)({
   alignItems: "flex-start",
   gap: "40px",
   width: "100%",
+  height: "100%",
 });
 
 export const StyledContent = styled(Box)({
@@ -14,19 +17,17 @@ export const StyledContent = styled(Box)({
   flexDirection: "row",
   width: "100%",
   gap: "40px",
-  alignContent: "flex-start",
-});
-
-export const StyledActions = styled(Box)({
-  display: "flex",
-  flexDirection: "column",
-  gap: "20px",
-  width: "70vw",
+  height: "100%",
   alignItems: "flex-start",
 });
 
-export const StyledActionsRow = styled(Box)({
-  display: "flex",
-  flexDirection: "row",
-  gap: "20px",
+export const StyledMenu = styled(Card)({
+  backgroundColor: theme.palette.primary[50],
+  minWidth: "15vw",
+  border: "none",
+});
+
+export const StyledLink = styled(Link)({
+  textDecoration: "none",
+  color: theme.palette.primary.main,
 });
