+++ src/pages/AdminUsers/AdminUsers.styles.tsx
@@ -1,0 +1,10 @@
+import styled from "@emotion/styled";
+import { Box } from "@mui/material";
+
+export const StyledContent = styled(Box)({
+  display: "flex",
+  flexDirection: "row",
+  width: "100%",
+  gap: "40px",
+  alignItems: "flex-start",
+});
