+++ src/pages/index.ts
@@ -1,5 +1,8 @@
 export * from "./Account";
 export * from "./Admin";
+export * from "./AdminHours";
+export * from "./AdminReports";
+export * from "./AdminUsers";
 export * from "./Auth";
 export * from "./Events";
 export * from "./Home";
